# docs/swagger.yaml
basePath: /
definitions:
  domain.Comment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  domain.Photo:
    properties:
      caption:
        type: string
      comment:
        items:
          $ref: '#/definitions/domain.Comment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  domain.SocialMedia:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  domain.SocialMediaAdd:
    properties:
      name:
        type: string
      social_media_url:
        type: string
    required:
    - name
    - social_media_url
    type: object
  domain.SocialMediaAddResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  domain.SocialMediaUpdateData:
    properties:
      name:
        type: string
      social_media_url:
        type: string
    required:
    - name
    - social_media_url
    type: object
  domain.SocialMediaUpdateDataResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  domain.SocialMediaWithUserData:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/domain.User'
      user_id:
        type: integer
    type: object
  domain.User:
    properties:
      age:
        type: integer
      comments:
        items:
          $ref: '#/definitions/domain.Comment'
        type: array
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      photo:
        $ref: '#/definitions/domain.Photo'
      social_media:
        $ref: '#/definitions/domain.SocialMedia'
      updated_at:
        type: string
      username:
        type: string
    type: object
  domain.UserDeleteResponse:
    properties:
      message:
        type: string
    type: object
  domain.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  domain.UserLoginResponse:
    properties:
      token:
        type: string
    type: object
  domain.UserRegister:
    properties:
      age:
        minimum: 8
        type: integer
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - age
    - email
    - password
    - username
    type: object
  domain.UserRegisterResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  domain.UserUpdateData:
    properties:
      email:
        type: string
      username:
        type: string
    required:
    - email
    - username
    type: object
  domain.UserUpdateDataResponse:
    properties:
      email:
        type: string
      username:
        type: string
    type: object
  middleware.Error:
    properties:
      message:
        type: string
    type: object
host: localhost:8001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample gin server for managing user's photos and social medias
    with Authentication & Authorization.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Photos & Social Meida Management API
  version: "1.0"
paths:
...
  /socialmedias:
    get:
      parameters:
      - description: Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SocialMediaWithUserData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Error'
      summary: Get all social medias
      tags:
      - User's Social Media
    post:
      consumes:
      - application/json
      parameters:
      - description: Access Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Social Media Data
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/domain.SocialMediaAdd'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.SocialMediaAddResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Error'
      summary: Add user's social media
      tags:
      - User's Social Media
  /socialmedias/{socialMediaId}:
    delete:
      parameters:
      - description: Access Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Social Media ID
        in: path
        name: socialMediaId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SocialMediaUpdateDataResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Error'
      summary: Delete existing user's social media
      tags:
      - User's Social Media
    put:
      consumes:
      - application/json
      parameters:
      - description: Access Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Social Media ID
        in: path
        name: socialMediaId
        required: true
        type: integer
      - description: Social Media Data
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/domain.SocialMediaUpdateData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SocialMediaUpdateDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Error'
      summary: Update existing user's social media
      tags:
      - User's Social Media
  /users:
    delete:
      parameters:
      - description: Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserDeleteResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Error'
      summary: Delete existing user account
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: Access Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.UserUpdateData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserUpdateDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Error'
      summary: Update user account data
      tags:
      - Users
  /users/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Error'
      summary: Login with existing user account
      tags:
      - Users
  /users/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.UserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.UserRegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Error'
      summary: Register a new user account
      tags:
      - Users
swagger: "2.0"